+++ classes/helper/helper.php
@@ -52,10 +52,9 @@ class Helper {
 			return $img_ids;
 		}
 
-		global $wpdb;
 		foreach ( $urls[1] as $url ) {
 			// Check if retrieved media from href really exists for the current site
-			$attachment_id = $wpdb->get_col( $wpdb->prepare( "SELECT ID FROM {$wpdb->posts} WHERE guid='%s';", $url ) );
+			$attachment_id = self::get_attachment_id_from_url( $url );
 			if ( empty( $attachment_id ) ) {
 				continue;
 			}
@@ -63,6 +62,44 @@ class Helper {
 		}
 
 		return $img_ids;
+	}
+
+	/**
+	 * Transform full URL in attachment ID
+	 *
+	 * @param string $attachment_url
+	 *
+	 * @link   https://philipnewcomer.net/2012/11/get-the-attachment-id-from-an-image-url-in-wordpress/
+	 *
+	 * @since  future
+	 * @author Amaury Balmer
+	 *
+	 * @return int
+	 */
+	static function get_attachment_id_from_url( string $attachment_url ) {
+		global $wpdb;
+
+		// If there is no url, return.
+		if ( '' == $attachment_url ) {
+			return 0;
+		}
+
+		// Get the upload directory paths
+		$upload_dir_paths = wp_upload_dir();
+
+		// Make sure the upload path base directory exists in the attachment URL, to verify that we're working with a media library image
+		if ( false !== strpos( $attachment_url, $upload_dir_paths['baseurl'] ) ) {
+			// If this is the URL of an auto-generated thumbnail, get the URL of the original image
+			$attachment_url = preg_replace( '/-\d+x\d+(?=\.(jpg|jpeg|png|gif)$)/i', '', $attachment_url );
+
+			// Remove the upload path base directory from the attachment URL
+			$attachment_url = str_replace( $upload_dir_paths['baseurl'] . '/', '', $attachment_url );
+
+			// Finally, run a custom database query to get the attachment ID from the modified attachment URL
+			return (int) $wpdb->get_var( $wpdb->prepare( "SELECT wposts.ID FROM $wpdb->posts wposts, $wpdb->postmeta wpostmeta WHERE wposts.ID = wpostmeta.post_id AND wpostmeta.meta_key = '_wp_attached_file' AND wpostmeta.meta_value = '%s' AND wposts.post_type = 'attachment'", $attachment_url ) );
+		}
+
+		return 0;
 	}
 
 	/**