+++ classes/helpers.php
@@ -22,8 +22,11 @@ class Helpers {
 			case 'acf' :
 				$label = _x( 'Advanced Custom Fields', 'Label for humanizing object types', 'bea-media-analytics' );
 				break;
+			case 'acf-option' :
+				$label = _x( 'Advanced Custom Fields (settings)', 'Label for humanizing object types', 'bea-media-analytics' );
+				break;
 			default :
-				$label = '';
+				$label = apply_filters( 'bea.media_analytics.helpers.humanize_object_type', '', $type );
 				break;
 		}