+++ classes/wp-cli/index-site.php
@@ -1,5 +1,6 @@
 <?php namespace BEA\Media_Analytics\WP_Cli;
 
+use BEA\Media_Analytics\Admin\Option;
 use BEA\Media_Analytics\Admin\Post;
 
 class Index_Site extends \WP_CLI_Command {
@@ -15,27 +16,57 @@ class Index_Site extends \WP_CLI_Command {
 	 *
 	 * @synopsis
 	 */
-	function index_site() {
+	public function index_site() {
 		plugins_loaded_bea_media_analytics_plugin();
 
+		$i = 0;
+
+		// Posts
 		$contents_q = new \WP_Query( [
 			'no_found_rows' => true,
 			'nopaging'      => true,
 			'post_type'     => 'any',
 			'post_status'   => 'any'
 		] );
-		if ( ! $contents_q->have_posts() ) {
-			\WP_CLI::error( sprintf( 'No content to index.' ) );
+    
+		if ( $contents_q->have_posts() ) {
+			$progress = \WP_CLI\Utils\make_progress_bar( sprintf( 'Indexing %s posts for blog_id %s', $contents_q->found_posts, get_current_blog_id() ), $contents_q->found_posts );
+		  foreach ( $contents_q->posts as $post ) {
+				$i ++;
 
-			return;
+				Post::index_post( $post->ID, $post, true );
+				$progress->tick();
+			}
+
+			$progress->finish();
+		} else {
+			\WP_CLI::warning( 'No post to index.' );
 		}
 
-		$progress = \WP_CLI\Utils\make_progress_bar( sprintf( 'Indexing %s contents for blog_id %s', $contents_q->post_count, get_current_blog_id() ), $contents_q->post_count );
-		foreach ( $contents_q->posts as $post ) {
-			Post::index_post( $post->ID, $post, true );
-			$progress->tick();
+		// ACF
+		if ( function_exists( 'acf_options_page' ) ) {
+			$pages = acf_options_page()->get_pages();
+
+			if ( empty( $pages ) ) {
+				\WP_CLI::warning( 'No settings page to index.' );
+			} else {
+				$total = count( $pages );
+
+				$progress = \WP_CLI\Utils\make_progress_bar( sprintf( 'Loop on settings page for blog id %d', get_current_blog_id() ), $total );
+				foreach ( $pages as $page ) {
+					$i ++;
+
+					Option::index_page_option( $page['menu_slug'] );
+
+					$progress->tick();
+				}
+
+				$progress->finish();
+			}
 		}
 
-		$progress->finish();
+		do_action_ref_array( 'bea.media_analytics.cli.index_site', array( &$i ) );
+
+		\WP_CLI::success( sprintf( '%s indexed contents for blog id : %d !', $i, get_current_blog_id() ) );
 	}
 }