+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fix wp-cli command registering.
 * Improve speed of wp-cli.
 * Add wp-cli progress bar for indexation command.
+* Add Component Field support.
+* Add Image Map Pro support.
 
 ## 2.0.3 - 04 Feb 2018
 * Go to WP.Org.